Converted all code to four spaces, not tabs.
This is my personal preference and I understand this may be controversial
going forward, but right now this project is a personal hobby, and in my
experience, tabs, while useful and sensible, are often used incorrectly.
It is unfortunate, but the simplest way to avoid that is to ban them.



git-svn-id: http://svgweb.googlecode.com/svn/trunk@5 4490c68e-677f-11de-b8ef-473237edd27e