Improved the default scaling mode based on the way SVG works.
Replaced the viewbox implementation. If you set the objectWidth
and objectHeight to the SVG width and height, the image
should look the same as a browser would render it.

Implemented a scripting interface, which currently works 
for simple object creation and attribute setting.

Implemented svg script parsing and implemented partial proxies
for svg scripting objects using browser javascript.

Increased the version to 0.7 due to new API, scale mode, and
start of scripting support.


git-svn-id: http://svgweb.googlecode.com/svn/trunk@39 4490c68e-677f-11de-b8ef-473237edd27e