Issue 179; Fix uncaught exceptions found using the flash 9 debug activeX control:
Fixed problem with getStyleOrAttr returning null for
opacity, when a default value is not provided. This
changes an assumption that the caller must always provide
a default value. If the caller is someone without such
knowledge (SMIL routines) then getStyleOrAttr should
return the appropriate default when the caller does not
provide one (by specifying null for the defaultValue parameter).

Fixed handlers for media events for audio and video elements.

Fixed problem with text characters trying to register themselves
as glyphs, when in fact they are clones of one master glyph
and that is the only glyph that should register with the font.

Fixed problem with SMIL routine where there was no
active interval but the code assumed there was one.



git-svn-id: http://svgweb.googlecode.com/svn/trunk@643 4490c68e-677f-11de-b8ef-473237edd27e